Chapter 1 finished

+++ chap-01/4-link-errors/3-sub.cpp
@@ -1,6 +1,6 @@
 #include "3-sub.hpp"
 
-float sub(float a, float b)
+float sub(int a, int b)
 {
     return a - b;
 }
